Add Draft Swipe
Added some basic swipe operations.
There are still some issues with it. Two Main issues:
1. Once you swipe to left or right, there is no way to go neutralize
it.
2. Once you swipe the bar, the second time when you click on wherever
on the bar, it will trigger the operation.

Good News:
It allow us to do different operations based on user operation. i.e, we
can control the behavior of the user when they swipe to left versus to
the right.